requirements.txt
Django==3.2
djangorestframework>=3.12.4,<3.13
psycopg2-binary>=2.9,<2.9.3
drf-spectacular>=0.15.1,<0.16
drf-jwt
pillow